Fix bug related to client page submission filter
Clicking Apply Filters button triggered the update of the submission selection, which changed the selected submission to the default. This was a not expected behaviour and needed to be fixed. To do so, we are now using a fixed list of choices that come from the dm object created when the app is launched. It is updated in the server. The observeEvent will trigger once because list_quarters is not reactive.