Change ethnicity filter
Ethnicity column contains comma separated values because a client can have more than one ethnicity (for example: White,Asian). Ethnicity input is a character vector of ethnicities (for example: c("Asian", "White", "Black ..."). In order to select the clients that have selected ethnicities I used stringr::str_detect using as a pattern the collapsed "|" separated inputs.

However, this approach comes with a drawback. Let's suppose that we selected in the input "White" and "Asian" and un-selected the rest. If a person is "Black" and "Asian", because we have "Asian" selected, it will be present (even though we have not selected "Black").

I imagine there is a way of removing un-selected ethnicities using some sort of complement of input$ethnicity and stringr::str_detect( negate = TRUE ). If we go for this approach, the filter would ONLY show clients with selected ethnicities that don't have un-selected ethnicities. In our previous example, our "White" and "Asian" client would not be selected, because even though "Asian" is selected, "Black" is not.